Make it work for multiple zones at the same time
Fixes an error that would occur when saving two zones at the same time:
Error:
   0: Cannot join child thread
   1: Child threw an error
   2: While watching zone `example.com`
   3: Cannot sync state to database
   4: Cannot INSERT into zones table
   5: error returned from database: (code: 5) database is locked
   6: (code: 5) database is locked

The fix should not be necessary, this should just work out of the box IMO. Or maybe something like `BEGIN IMMEDIATE` would be needed in sqlx but that does not seem to be implemented.
I don't know how else to fix this.

Also add a test to make sure it won't break in the future.