In test change group of directory again
This has no effect on the test itself but the code may be used as an example where it does matter.
Since the systemd unit runs with DynamicUser=true, the group ID of the directory may change over time, so reset it.
This reverts part of b913fc2f7be8395e85e7f63fc639f6518fe2e4e0.