Don't use a separate thread for each zone
Tokio still uses multiple threads, so the performance should be comparable (not that I need it).
The original idea was to separate all the zones from each other as much as possible.
But I no longer want the additional complexity and overhead of multiple threads when async alone also works.
One zone may now theoretically have more influence over the performance of another zone but since I don't have thousands of zones, this is fine.
This simplifies the code quite a bit.

I couldn't figure out how to name Tokio threads without unstable features, so to preserve the information from which zone a log message comes, I manually added this information everywhere. I hope I didn't forget any log messages. This is a new (minor) risk and a downside of this approach.
Stack traces are nicer though since the part of the code spawning the threads no longer exists and so isn't visible in the log messages anymore.